requirements.txt
aiogram
sqlalchemy
python-dotenv
alembic
pydantic